Add project options menu to cards
