+++ app/scripts/stores/mainStore.js
@@ -245,9 +245,10 @@ export class MainStore {
             } else {
                 this.projects = [...this.projects, ...results];
             }
+            const userId = authStore.currentUser.id !== undefined ? authStore.currentUser.id : this.currentUser.id !== undefined ? this.currentUser.id : null;
             this.projects.forEach((p) => {
-                this.getAllProjectPermissions(p.id, authStore.currentUser.id)
-            })
+                userId !== null ? this.getAllProjectPermissions(p.id, authStore.currentUser.id) : null;
+            });
             this.responseHeaders = headers;
             this.loading = false;
         }).catch(ex => this.handleErrors(ex))
@@ -311,6 +312,8 @@ export class MainStore {
             .then((json) => {
                 this.addToast('Project Updated');
                 this.project = json;
+                let index = this.projects.findIndex((p) => p.id === id);
+                this.projects.splice(index, 1, json);
             }).catch((ex) => {
             this.addToast('Project Update Failed');
             this.handleErrors(ex)
@@ -323,7 +326,7 @@ export class MainStore {
             .then(response => {})
             .then(() => {
                 this.addToast('Project Deleted');
-                BaseUtils.removeObjByKey(this.projects, {key: 'id', value: id})
+                this.projects = this.projects.filter(p => p.id !== id);
             }).catch((ex) => {
             this.addToast('Project Delete Failed');
             this.handleErrors(ex)
@@ -527,6 +530,15 @@ export class MainStore {
                     this.isListItem = isListItem;
                 }).catch(ex => this.handleErrors(ex))
         }
+    }
+
+    @action setSelectedProject(id) {
+        this.transportLayer.getProjectDetails(id)
+            .then(this.checkResponse)
+            .then(response => response.json())
+            .then((json) => {
+                this.project = json;
+            }).catch(ex => this.handleErrors(ex))
     }
 
     @action getObjectMetadata(id, kind) {
